+++ CHANGELOG.md
@@ -76,6 +76,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Ambient noise now gets filtered when underwater.
 - Renamed Site Savannah Pit to Savannah Town
 - Replaced building Savannah Pit with Savannah Town Airship Dock
+- Server authoritative physics forces now persist across server restarts.
 
 ### Removed